+++ endpoints/azure/src/lib.rs
@@ -1,7 +1,6 @@
 mod host;
 // TODO: Attempt to make the modules non-pub
 pub mod synthesize;
-pub mod synthesize_service;
 pub mod transcribe;
 
 pub use host::*;
